ml-engine/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
pymongo==4.6.1
mysql-connector-python==8.2.0
